[rpc] fix assertion errors in record_function scope when running profiler with rpc
Fix for GH issue https://github.com/pytorch/pytorch/issues/32517. The
issue is that the `record_function` decorator uses nested RecordFunction
scopes that are handled by appropriately setting `parent_` and a thread_local
pointer to the current RecordFunction. This was not properly set with RPC
profiling, since `RecordFunction::end()` is called from a different thread so
the original thread_local object is invalid.

This fixes the issue by saving a pointer to the `thread_local` pointer, and
then properly setting the pointer to the parent_ RecordFunction.

There are still some issues with the profiler output with RPC and the
`record_function` decorator. Namely if we do
```
with record_function("foo"):
   rpc.rpc_sync(...)
```

we'd expect the sum of rpc_sync and foo() to be 100% of CPU time, like it is
for other ops, but this is not currently the case. We should fix this.

Differential Revision: [D19549107](https://our.internmc.facebook.com/intern/diff/D19549107/)

ghstack-source-id: 97142437
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32566