[jit] Delete the ErrorReport default constructor
An error report without a SourceRange context is bad, because it doesn't
tell the user where something happened. Delete the default constructor
to make it harder to create errors like this (you can still use a fake
SourceRange if you absolutely need to).

Also clean up the only case where the default constructor was used.

ghstack-source-id: 48ae96994552200f7c3e4ef987a6c2f6d7478cac
Pull Request resolved: https://github.com/pytorch/pytorch/pull/32879