[profiler] remove redundant assert in record_function_ops
This removes a redundant assert statement in `record_function_ops`. In
the else branch in question, we are guaranteed to have `current == &rec`, so
this assert will never fire.

Although, maybe we should add an assert failure when `current == &rec` since it
seems that `current` should always be profiler::record_function_exit.

Differential Revision: [D19849145](https://our.internmc.facebook.com/intern/diff/D19849145/)

ghstack-source-id: 98145667
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33225