Add mechanism to pass a number of workers to cpp extensions
Fixes #33091

This PR lets users control the number of workers that cpp extensions
uses through the environment variable `MAX_JOBS`. If the environment
variable is a non-negative integer we use that many threads; otherwise,
ninja falls back to the default.

I chose to use the name `MAX_JOBS` because we use it in PyTorch already
to control the number of workers PyTorch builds with. There is a risk
that users of cpp extensions already have `MAX_JOBS` set but we are
hoping that that risk is small and/or it means semantically the same
thing.

Test Plan:
- tested locally

ghstack-source-id: 85fa57c53d7d0134f7b18bc20875823501d7ce21
Pull Request resolved: https://github.com/pytorch/pytorch/pull/33346